move handleError to utils
